feat: Added interface to represent a Resource
Having a resource be represented by an interface will allow for the
abstraction of that resource.

With that, we can now easily change the implementation of what a
resource should be without having to change it across the code.

This interface creates a contract on the behaviour of the resource,
making it easy to work with.