feat: Added ability to decorate/proxy a Resource
To use this, simply extend this class and `override` it's
implementation recalling the equivelent method of `super` when
retrieving the original behaviour.

Not only a basic Resource can be decorated/proxied. A decorator can also
be decorated by other decorators (same is true for a proxy).

The main difference between a decorator and a proxy is that a proxy can
break the behavioural chain whereas the decorator doesn't, but all this
under the same API. The way the end user uses is what will tell if it's
one or another.

Based on Decorator and Proxy Pattern.
https://refactoring.guru/design-patterns/decorator
https://refactoring.guru/design-patterns/proxy