feat: Added interface to represent a Handler
Having a handler be represented by an interface will allow for the
abstrction of that handler.

With that, we can now easily change the implementation of what a handler
should be, without having to change it across the code.

This interface creates a contract on the behaviour of the handler,
making it easy to work with.

Based on Chain of Responsibility Pattern.
https://refactoring.guru/design-patterns/chain-of-responsibility